refactor(error): add an abstract message error service
